+++ src/notes/dto/note.dto.ts
@@ -1,5 +1,5 @@
 import { IsIn, IsString, MaxLength, MinLength } from 'class-validator';
-import { Category } from '../types/Category';
+import { Category } from '../../types/Category';
 
 const categories = ['task', 'quote', 'idea', 'thought'];
 
@@ -8,8 +8,10 @@ export class NoteDto {
 	@MinLength(1)
 	@MaxLength(64)
 	name: string;
+
 	@IsIn(categories)
 	category: Category;
+
 	@IsString()
 	@MinLength(1)
 	content: string;
